# Conflicts:
#	README.md
#	api/data/api/handlers/SpiceBeanHandler.php
#	api/include/Localization/Localization.php
#	api/include/Soap/SpiceSoap.php
#	api/include/Soap/SpiceSoapRegistry.php
#	api/include/Soap/SpiceSoapServiceImpl.php
#	api/include/Soap/SpiceSoapServiceUtil.php
#	api/include/Soap/soap.php
#	api/include/SpiceDemoData/SpiceDemoDataGenerator.php
#	api/include/SugarObjects/LanguageManager.php
#	api/include/authentication/AuthenticationController.php
#	api/include/authentication/api/controllers/AuthenticateController.php
#	api/modules/CampaignTasks/CampaignTask.php
#	api/modules/CampaignTasks/api/controllers/CampaignTasksController.php
#	api/vendor/nusoap/nusoap.php
#	app/admincomponents.js
#	app/cleverreach.js
#	app/default-src_include_spicetimeline_spicetimeline.js
#	app/default-src_modules_activities_components_activitytimelineadditem.js
#	app/default-src_modules_activities_moduleactivities.js
#	app/default-src_modules_mediafiles_modulemediafiles.js
#	app/default-src_modules_outputtemplates_components_objectactionoutputbeanmodal.js
#	app/default-src_modules_reports_modulereports.js
#	app/default-src_modules_reportsdesigner_modulereportsdesigner.js
#	app/dialogmail.js
#	app/evalanchemodule.js
#	app/exchange.js
#	app/gsuitesettings.js
#	app/mailchimp.js
#	app/main.js
#	app/moduleaccounts.js
#	app/moduleacl.js
#	app/moduleaclterritories.js
#	app/moduleagreements.js
#	app/modulealcatel.js
#	app/moduleasterisk.js
#	app/modulebonusprograms.js
#	app/modulecalendar.js
#	app/modulecampaigns.js
#	app/modulecatalogs.js
#	app/modulechat.js
#	app/modulecontacts.js
#	app/modulecurrencies.js
#	app/moduledashboard.js
#	app/moduledeployment.js
#	app/moduledocuments.js
#	app/moduleduns.js
#	app/moduleemails.js
#	app/modulefive9.js
#	app/modulefolders.js
#	app/modulegoogleapi.js
#	app/moduleholidaycalendars.js
#	app/modulehome.js
#	app/moduleknowledge.js
#	app/moduleleads.js
#	app/modulemailboxes.js
#	app/moduleopportunities.js
#	app/moduleorgunits.js
#	app/moduleoutputtemplates.js
#	app/modulepipl.js
#	app/modulepotentials.js
#	app/modulepriceconditions.js
#	app/moduleproducts.js
#	app/moduleprojects.js
#	app/moduleprospectlists.js
#	app/modulequestionnaires.js
#	app/modulereportsdesignermore.js
#	app/modulereportsmore.js
#	app/modulesalesdocs.js
#	app/modulesalesplanning.js
#	app/modulesap.js
#	app/modulesapidocs.js
#	app/modulescrum.js
#	app/moduleservicecalendars.js
#	app/modulespiceimports.js
#	app/modulespicepagebuilder.js
#	app/modulestarface.js
#	app/modulesystemcalendars.js
#	app/modulesystemtenants.js
#	app/moduletelephony.js
#	app/moduletelesales.js
#	app/moduletotpauthentication.js
#	app/moduleusers.js
#	app/moduleworkflow.js
#	app/portalcomponents.js
#	app/quilleditor.js
#	app/runtime.js
#	app/scripts.js
#	app/sendinbluemodule.js
#	app/servicecomponents.js
#	app/spiceattachments.js
#	app/spicebcardreader.js
#	app/spicecharts.js
#	app/spicefavorites.js
#	app/spiceimporter.js
#	app/spicemap.js
#	app/spicemath.js
#	app/spicemerge.js
#	app/spicenotes.js
#	app/spicepath.js
#	app/spicetexts.js
#	app/spicetimestream.js
#	app/styles.css
#	app/vendor.js
#	app/workbench.js
#	environments/environment.prod.ts
#	gsuite.html
#	index.html
#	outlook.html
#	outlookcrm.html
#	src/objectfields/components/fieldparent.ts

+++ api/include/SpiceUI/SpiceUIConfLoader.php
@@ -84,6 +84,31 @@ class SpiceUIConfLoader
         'sysuirolemodules',
         'sysuiroles',
         'sysuiroutes',
+        'sysexchangemappingsegments',
+        'sysexchangemappingsegmentitems',
+        'sysexchangemappingmodules',
+        'sysmsgraphmappingsegments',
+        'sysmsgraphmappingsegmentitems',
+        'sysmsgraphmappingmodules',
+    ];
+
+    /**
+     * specific records from tables without package column shall be inserted, NOT updated, NOT deleted
+     * @var string[]
+     */
+    private $insertOnlyTables = [
+        'nodata',
+        'sysfts',
+        'sysnumberranges',
+        'sysnumberrangeallocation',
+        'syssalesdocnumberranges',
+        'syssalesdoctypes',
+        'syssalesdoctypesflow',
+        'syssalesdoctypesitemtypes',
+        'syscategorytrees',
+        'syscategorytreelinks',
+        'schedulerjobtasks',
+        'schedulerjobs'
     ];
 
     /**
@@ -184,12 +209,29 @@ class SpiceUIConfLoader
             throw new Exception($errormsg);
         }
 
-        $tables = array_keys($response);
+        // gather tables and all the record IDs
+        foreach ($response as $conftable => $conf){
+            foreach($conf as $recordId => $recordData){
+                $tables[$conftable][] = $recordId;
+            }
+        }
 
-        foreach ($tables as $conftable){
-            $delWhere = ['package' => $package];
-            if(!$db->deleteQuery($conftable, $delWhere)){
-                LoggerManager::getLogger()->fatal('error deleting package {$package}  '.$db->lastError());
+        foreach ($tables as $conftable => $recordIds){
+            if(in_array($conftable, $this->insertOnlyTables)){
+                continue;
+            }
+            try {
+                $delWhere = ['package' => $package];
+                $db->deleteQuery($conftable, $delWhere);
+            } catch (Exception $e){
+                // just go on -  the table just has no package column
+            }
+
+            foreach($recordIds as $recordId){
+                $delWhere = ['id' => $recordId];
+                if (!$db->deleteQuery($conftable, $delWhere)) {
+                    LoggerManager::getLogger()->fatal('error deleting package {$package}  ' . $db->lastError());
+                }
             }
         }
     }
@@ -336,11 +378,28 @@ class SpiceUIConfLoader
                     foreach ($decodeData as $key => $value) {
                         $decodeData[$key] = (is_null($value) || $value === "" ? NULL :  $value);
                     }
+
+                    // set the flag to check on insert
+                    $skipInsert = false;
+
                     //delete before insert
-                    $delWhere = ['id' => $decodeData['id']];
-                    if(!$db->deleteQuery($tb, $delWhere)){
-                        LoggerManager::getLogger()->fatal("error deleting entry {$decodeData['id']} ".$db->lastError());
+                    if(!in_array($tb, $this->insertOnlyTables)){
+                        $delWhere = ['id' => $decodeData['id']];
+                        if(!$db->deleteQuery($tb, $delWhere)){
+                            LoggerManager::getLogger()->fatal("error deleting $tb entry {$decodeData['id']} ".$db->lastError());
+                        }
+                    } else{
+                        // check if record is present
+                        if($dbResRow = $db->getOne("select * from $tb where id='{$decodeData['id']}'")){
+                            $skipInsert = true;
+                        }
                     }
+
+                    // skip insert if this is a record we should keep
+                    if($skipInsert) {
+                        continue;
+                    }
+
                     //run insert
 //                if($tb == 'email_templates'){
 //                    file_put_contents('spicecrm.log', 'dict email_templates '.print_r(SpiceDictionaryHandler::getInstance()->dictionary['EmailTemplate'], true)."\n", FILE_APPEND);
